tools/essence-feature-usage-stats/requirements.txt
blinker==1.6.3
click==8.1.7
Flask==3.0.0
gitdb==4.0.10
GitPython==3.1.37
itsdangerous==2.1.2
Jinja2==3.1.2
markdown-it-py==3.0.0
MarkupSafe==2.1.3
mdurl==0.1.2
Pygments==2.16.1
python-dotenv==1.0.0
rich==13.6.0
smmap==5.0.1
tqdm==4.66.1
Werkzeug==3.0.0
